+++ packages/rule/int.ts
@@ -1,6 +1,0 @@
-/**
- * Isotropic non-totalistic rule
- * @module
- */
-export { intConditionArray } from "./lib/int.ts";
-export { intModifiers } from "./lib/int/int-condition.ts";