+++ packages/rule/lib/grid/mod.ts
@@ -7,11 +7,15 @@ export type GridParameter = {
    */
   size: {
     /**
-     * 0 is infinite
+     * Width of the grid.
+     *
+     * 0 represents infinite width.
      */
     width: number;
     /**
-     * 0 is infinite
+     * Height of the grid.
+     *
+     * 0 represents infinite height.
      */
     height: number;
   };
@@ -236,6 +240,14 @@ export function stringifyGridParameterWithColon(
 
 export function stringifyGridParameter(gridParameter: GridParameter): string {
   const size = gridParameter.size;
+  if (!Number.isInteger(size.width) || !Number.isInteger(size.height)) {
+    throw new Error("size is not a integer");
+  }
+
+  if (size.width < 0 || size.height < 0) {
+    throw new Error("size is less than 0");
+  }
+
   switch (gridParameter.topology.type) {
     case "P": {
       return `P${size.width},${size.height}`;
@@ -283,7 +295,11 @@ export function stringifyGridParameter(gridParameter: GridParameter): string {
 function encodeShift(amount: number | null | undefined) {
   if (amount == null) {
     return "";
-  } else if (amount >= 0) {
+  } else if (Number.isNaN(amount)) {
+    throw new Error("shift is NaN");
+  } else if (amount === 0) {
+    return "";
+  } else if (amount > 0) {
     return "+" + amount;
   } else {
     return "-" + (-amount);